+++ lib/lightning/extensions/usage_limiting.ex
@@ -2,16 +2,21 @@ defmodule Lightning.Extensions.UsageLimiting do
   @moduledoc """
   Rate limiting for Lightning workloads that depends on Runtime.
   """
-  @type error_reason :: :too_many_runs | :runs_hard_limit | :unknown_project
+  @type error_reason ::
+          :too_many_runs
+          | :runs_hard_limit
+          | :unknown_project
   @type message :: Lightning.Extensions.Message.t()
+  @type error :: {:error, error_reason(), message()}
 
   defmodule Action do
     @moduledoc false
     @type t :: %__MODULE__{
-            type: :new_run | :new_workflow
+            type: :new_run | :new_workflow,
+            amount: pos_integer()
           }
 
-    defstruct [:type]
+    defstruct type: nil, amount: 1
   end
 
   defmodule Context do
@@ -25,8 +30,8 @@ defmodule Lightning.Extensions.UsageLimiting do
   end
 
   @callback check_limits(context :: Context.t()) ::
-              :ok | {:error, error_reason(), message()}
+              :ok | error()
 
   @callback limit_action(action :: Action.t(), context :: Context.t()) ::
-              :ok | {:error, error_reason(), message()}
+              :ok | error()
 end